refactor: package name from sms_otp changed to sms_service
