+++ pkg/sms_service/sms_service.go
@@ -1,4 +1,4 @@
-package sms_otp
+package sms_service
 
 import (
 	"bytes"
@@ -12,16 +12,16 @@ import (
 
 const TEMPLATE_FORMAT = "txt"
 
-type SMSOtp struct {
+type SmsService struct {
 	configs       *config.SMS
 	templatesPath string
 }
 
-func NewSMSOtpService(configs *config.SMS, templatesPath string) *SMSOtp {
-	return &SMSOtp{configs, templatesPath}
+func NewSmsService(configs *config.SMS, templatesPath string) *SmsService {
+	return &SmsService{configs, templatesPath}
 }
 
-func (s *SMSOtp) SendSMS(msg string, receivers []string) error {
+func (s *SmsService) SendSMS(msg string, receivers []string) error {
 	if config.ENV == "devel" {
 		fmt.Println("------ SMS Sent ------")
 		fmt.Printf("msg: \n%s\n", strings.TrimSpace(msg))
@@ -37,7 +37,7 @@ func (s *SMSOtp) SendSMS(msg string, receivers []string) error {
 }
 
 // Parses and returns the template.
-func (s *SMSOtp) Template(name string, args interface{}) (string, error) {
+func (s *SmsService) Template(name string, args interface{}) (string, error) {
 	filename := fmt.Sprintf("%s/%s.%s", s.templatesPath, name, TEMPLATE_FORMAT)
 
 	fileData, readErr := os.ReadFile(filename)
